+++ CHANGELOG.md
@@ -9,8 +9,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Added
 
+## Fixed
+
+## Changed
+
+## Removed
+
+## [1.3.0] - 2023-05-07
+
+## Added
+
 -   Stop mid-air (ArrowDown)
 -   Key bindings (H, J, K, L)
+-   Message for mobile users
 
 ## Fixed
 
@@ -19,6 +30,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Changed
 
 -   Pixel deleting logic - now pixel is always just toggled
+-   Database removes players after them being inactive for 60 seconds (instead of midnight ritual)
+-   Out of focus -> auto refresh
 
 ## Removed
 
@@ -100,7 +113,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Online multiplayer environment with HTML Canvas and Firebase
 -   No tests, no for phone, also lag
 
-[unreleased]: https://github.com/naglissul/jsgd/compare/v1.2.0...HEAD
+[unreleased]: https://github.com/naglissul/jsgd/compare/v1.3.0...HEAD
+[1.3.0]: https://github.com/naglissul/jsgd/compare/v1.2.0...v1.3.0
 [1.2.0]: https://github.com/naglissul/jsgd/compare/v1.1.1...v1.2.0
 [1.1.1]: https://github.com/naglissul/jsgd/compare/v1.1.0...v1.1.1
 [1.1.0]: https://github.com/naglissul/jsgd/compare/v1.0.0...v1.1.0